+++ src/google-payment/index.js
@@ -1,14 +1,13 @@
 'use strict';
 /**
  * @module braintree-web/google-payment
- * @description A component to integrate with Google Pay.
+ * @description A component to integrate with Google Pay. The majority of the integration uses [Google's pay.js JavaScript file](https://payments.developers.google.com/js/apis/pay.js). The Braintree component generates the configuration object necessary for Google Pay to initiate the Payment Request and parse the returned data to retrieve the payment method nonce which is used to process the transaction on the server.
  *
  * **Note:** This component is currently in beta and the API may include breaking changes when upgrading. Please review the [Changelog](https://github.com/braintree/braintree-web/blob/master/CHANGELOG.md) for upgrade steps whenever you upgrade the version of braintree-web.
  * */
 
 var basicComponentVerification = require('../lib/basic-component-verification');
 var BraintreeError = require('../lib/braintree-error');
-var browserDetection = require('./browser-detection');
 var GooglePayment = require('./google-payment');
 var Promise = require('../lib/promise');
 var wrapPromise = require('@braintree/wrap-promise');
@@ -20,6 +19,104 @@ var VERSION = process.env.npm_package_version;
  * @param {object} options Creation options:
  * @param {Client} options.client A {@link Client} instance.
  * @param {callback} [callback] The second argument, `data`, is the {@link GooglePayment} instance. If no callback is provided, `create` returns a promise that resolves with the {@link GooglePayment} instance.
+ * @example <caption>Simple Example</caption>
+ * // include https://payments.developers.google.com/js/apis/pay.js in a script tag
+ * // on your page to load the `google.api.PaymentsClient` global object.
+ *
+ * var paymentButton = document.querySelector('#google-pay-button');
+ * var paymentsClient = new google.payments.api.PaymentsClient({
+ *   environment: 'TEST' // or 'PRODUCTION'
+ * });
+ *
+ * braintree.client.create({
+ *   authorization: 'tokenization-key-or-client-token'
+ * }).then(function (clientInstance) {
+ *   return braintree.googlePayment.create({
+ *     client: clientInstance
+ *   });
+ * }).then(function (googlePaymentInstance) {
+ *   paymentButton.addEventListener('click', function (event) {
+ *     var paymentDataRequest;
+ *
+ *     event.preventDefault();
+ *
+ *     paymentDataRequest = googlePaymentInstance.createPaymentDataRequest({
+ *       merchantId: 'your-merchant-id-from-google',
+ *       transactionInfo: {
+ *         currencyCode: 'USD',
+ *         totalPriceStatus: 'FINAL',
+ *         totalPrice: '100.00'
+ *       }
+ *     });
+ *
+ *     paymentsClient.loadPaymentData(paymentDataRequest).then(function (paymentData) {
+ *       var result = googlePaymentInstance.parseResponse(paymentData);
+ *
+ *       // send result.nonce to your server
+ *     }).catch(function (err) {
+ *       // handle err
+ *     });
+ *   });
+ * });
+ * @example <caption>Check Browser and Customer Compatibility</caption>
+ * function getPaymentsClient() {
+ *   return new google.payments.api.PaymentsClient({
+ *     environment: 'TEST' // or 'PRODUCTION'
+ *   });
+ * }
+ *
+ * function setupGooglePayButton(helper) {
+ *   var button = document.createElement('button');
+ *
+ *   button.id = 'google-pay';
+ *   button.appendChild(document.createTextNode('Google Pay'));
+ *   button.addEventListener('click', function (event) {
+ *     var paymentRequestData, paymentsClient;
+ *
+ *     event.preventDefault();
+ *
+ *     paymentDataRequest = helper.createPaymentDataRequest({
+ *       merchantId: 'your-merchant-id-from-google',
+ *       transactionInfo: {
+ *         currencyCode: 'USD',
+ *         totalPriceStatus: 'FINAL',
+ *         totalPrice: '100.00' // your amount
+ *       }
+ *     });
+ *     paymentsClient = getPaymentsClient();
+ *
+ *     paymentsClient.loadPaymentData(paymentDataRequest).then(function (paymentData) {
+ *       var result = helper.parseResponse(paymentData);
+ *
+ *       // send result.nonce to your server
+ *     }).catch(function (err) {
+ *       // handle errors
+ *     });
+ *   });
+ *
+ *   document.getElementById('container').appendChild(button);
+ * }
+ *
+ * braintree.client.create({
+ *   authorization: 'tokenization-key-or-client-token'
+ * }).then(function (clientInstance) {
+ *   return braintree.googlePayment.create({
+ *     client: clientInstance
+ *   });
+ * }).then(function (googlePaymentInstance) {
+ *   var paymentsClient = getPaymentsClient();
+ *
+ *   return paymentsClient.isReadyToPay({
+ *     allowedPaymentMethods: googlePaymentInstance.createPaymentDataRequest().allowedPaymentMethods
+ *   });
+ * }).then(function (response) {
+ *   if (response.result) {
+ *     setupGooglePayButton(googlePaymentInstance);
+ *   }
+ * }).catch(function (err) {
+ *   // handle setup errors
+ * });
+ *
  * @returns {Promise|void} Returns a promise if no callback is provided.
  */
 function create(options) {
@@ -27,8 +124,6 @@ function create(options) {
     name: 'Google Pay',
     client: options.client
   }).then(function () {
-    var googlePayment;
-
     if (!options.client.getConfiguration().gatewayConfiguration.androidPay) {
       return Promise.reject(new BraintreeError({
         type: BraintreeError.types.MERCHANT,
@@ -37,34 +132,12 @@ function create(options) {
       }));
     }
 
-    googlePayment = new GooglePayment(options);
-
-    return googlePayment.initialize();
+    return new GooglePayment(options);
   });
-}
-
-/**
- * @static
- * @function isSupported
- * @description Returns true if Google Pay is supported in this browser.
- * @example
- * if (braintree.googlePayment.isSupported()) {
- *    // Add Google Pay button to page and
- *    // initialize Google Pay component
- * } else {
- *    // Do not initialize Google Pay component
- * }
- * @returns {Boolean} Returns true if Google Pay supports this browser.
- */
-function isSupported() {
-  // TODO - We limit this to android chrome for now
-  // Once it works in Desktop Chrome, we can just check that it's Chrome
-  return Boolean(browserDetection.supportsPaymentRequestApi() && browserDetection.isAndroid());
 }
 
 module.exports = {
   create: wrapPromise(create),
-  isSupported: isSupported,
   /**
    * @description The current version of the SDK, i.e. `{@pkg version}`.
    * @type {string}